refactor: 파일 분리, 중복 제거
- YoutubeApp -> YoutubeMainApp, YoutubeModalApp으로 분리
- App과 dom에서 render 관련 함수 Render로 이동 및 분리
